+++ src/onegov.election_day/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds compatibility with Morepath 0.13.
+  [href]
+
 0.7.2 (2016-03-18)
 ~~~~~~~~~~~~~~~~~~~